+++ game-borderlands2/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - None Planned
 
+## [0.2.0] - 2025-09-19
+
+- Added SDK mod support.
+- Added support for BLCMM.
+- Added support for .blcm file mods.
+- Added installer for movies (.bik files).
+
 ## [0.1.0] - 2025-09-17
 
 - Initial release.